# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.UptimeMonitorLogSwagger:
    properties:
      checked_at:
        type: string
      error_message:
        type: string
      id:
        type: integer
      is_healthy:
        type: boolean
      response_time_ms:
        type: integer
      status_code:
        type: integer
      uptime_monitor_id:
        type: integer
    type: object
  controllers.UptimeMonitorSwagger:
    properties:
      code:
        type: string
      created_at:
        type: string
      healthy_checks:
        type: integer
      id:
        type: integer
      interval:
        type: integer
      is_active:
        type: boolean
      last_checked_at:
        type: string
      last_healthy_at:
        type: string
      last_unhealthy_at:
        type: string
      name:
        type: string
      next_check_at:
        type: string
      total_checks:
        type: integer
      unhealthy_checks:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  dto.StoreUptimeMonitorLogRequest:
    properties:
      checked_at:
        type: string
      error_message:
        type: string
      is_healthy:
        type: boolean
      response_time_ms:
        type: integer
      status_code:
        type: integer
      uptime_monitor_id:
        type: integer
    required:
    - checked_at
    - response_time_ms
    - status_code
    - uptime_monitor_id
    type: object
  dto.StoreUptimeMonitorRequest:
    properties:
      code:
        type: string
      interval:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      url:
        type: string
    required:
    - code
    - interval
    - name
    - url
    type: object
...
  dto.UpdateUptimeMonitorRequest:
    properties:
      code:
        type: string
      interval:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      url:
        type: string
    type: object
...
  utils.Meta:
    properties:
      current_page:
        type: integer
      has_more_pages:
        type: boolean
      items_on_page:
        type: integer
      per_page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        default: true
        type: boolean
    type: object
  utils.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@novaardiansyah.id
    name: API Support
    url: https://novaardiansyah.id
  description: This is an official Golang API documentation for NovaApp.
  license:
    name: MIT License
    url: https://github.com/novaardiansyah/golang-api/blob/main/LICENSE
  termsOfService: https://novaardiansyah.id/live/nova-app/terms-of-service
  title: Golang API
  version: "1.0"
paths:
...
  /uptime-monitor-logs:
    get:
      consumes:
      - application/json
      description: Get a paginated list of uptime monitor logs
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by monitor ID
        in: query
        name: monitor_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.UptimeMonitorLogSwagger'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List uptime monitor logs
      tags:
      - uptime_monitor_logs
    post:
      consumes:
      - application/json
      description: Store a new uptime monitor log
      parameters:
      - description: Log data
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/dto.StoreUptimeMonitorLogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.UptimeMonitorLogSwagger'
              type: object
      security:
      - BearerAuth: []
      summary: Store a new uptime monitor log
      tags:
      - uptime_monitor_logs
  /uptime-monitor-logs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing uptime monitor log
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete an uptime monitor log
      tags:
      - uptime_monitor_logs
    get:
      consumes:
      - application/json
      description: Get a single uptime monitor log by ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.UptimeMonitorLogSwagger'
              type: object
      security:
      - BearerAuth: []
      summary: Get uptime monitor log details
      tags:
      - uptime_monitor_logs
  /uptime-monitors:
    get:
      consumes:
      - application/json
      description: Get a paginated list of uptime monitors
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search by name or code
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.UptimeMonitorSwagger'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List uptime monitors
      tags:
      - uptime_monitors
    post:
      consumes:
      - application/json
      description: Store a new uptime monitor
      parameters:
      - description: Monitor data
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/dto.StoreUptimeMonitorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.UptimeMonitorSwagger'
              type: object
      security:
      - BearerAuth: []
      summary: Store a new uptime monitor
      tags:
      - uptime_monitors
  /uptime-monitors/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing uptime monitor
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete an uptime monitor
      tags:
      - uptime_monitors
    get:
      consumes:
      - application/json
      description: Get a single uptime monitor by ID
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.UptimeMonitorSwagger'
              type: object
      security:
      - BearerAuth: []
      summary: Get uptime monitor details
      tags:
      - uptime_monitors
    put:
      consumes:
      - application/json
      description: Update an existing uptime monitor
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Monitor data
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUptimeMonitorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.UptimeMonitorSwagger'
              type: object
      security:
      - BearerAuth: []
      summary: Update an uptime monitor
      tags:
      - uptime_monitors
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"